Introduced tabs for static and elastic agents
